requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# Computer Vision & Geospatial
opencv-python>=4.8.0
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.6.0
scikit-image>=0.21.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration & Logging
pyyaml>=6.0
omegaconf>=2.3.0
wandb>=0.15.0
tensorboard>=2.13.0

# Image Processing & Augmentation
albumentations>=1.3.0
Pillow>=10.0.0

# Clustering & Analysis
dbscan>=0.1.0
hdbscan>=0.8.0

# Utilities
tqdm>=4.65.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: CUDA support (uncomment if using GPU)
# torch-audio>=2.0.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118